+++ cfstoreviewer/urls.py
@@ -1,4 +1,4 @@
-"""cfstore_django URL Configuration
+"""cfstoreapp URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.1/topics/http/urls/
@@ -13,10 +13,11 @@ Including another URLconf
     1. Import the include() function: from django.urls import include, path
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
-from django.contrib import admin
-from django.urls import path,include
+from django.urls import path
+
+from . import views
 
 urlpatterns = [
-    path("admin/", admin.site.urls),
-    path('cfstoreviewer/', include('cfstoreviewer.urls'))
-]
+    path('', views.index, name='index'),
+    path('viewcollections',views.ls, name='allcollections'),
+    path('viewcollections/<str:page>',views.ls)]
